+++ canvas_modules/common-canvas/src/common-canvas/cc-panels.jsx
